Feat: "send others 액티비티 추가"
send others 액티비티 추가

Resolves: N/A
Ref: N/A
Related to: N/A
